First check-in of AVALON
AVALON is a Morphing Terrarium E350 inspired instrument which can read Cloud Terrarium E352 wavetable files. It can morph the wavetables in a XY 8x8 grid as well as Z morphing through all 64 waves. As an extra it can also morph between the XY and Z waves in a couple of different modes.